refactor(registry): absorb machine package
The purpose of the machine package was not clear, so move its
contents into the registry package.
