+++ registry/machine.go
@@ -1,11 +1,11 @@
 package registry
 
-type Interface struct {
-	Name string `json:"name"`
-	MTU int `json:"mtu"`
-	HardwareAddr string `json:"hardware_addr"`
-	Addrs []*Addr `json:"addrs"`
-}
+import (
+	"io/ioutil"
+	"strings"
+)
+
+const boot_id_path = "/proc/sys/kernel/random/boot_id"
 
 type Addr struct {
 	Addr string `json:"addr"`
@@ -13,5 +13,25 @@ type Addr struct {
 }
 
 type Machine struct {
-	Interfaces []*Interface `json:"interfaces"`
+	BootId string
+}
+
+func (m *Machine) String() string {
+	return m.BootId
+}
+
+func NewMachine(bootId string) (m *Machine) {
+	m = &Machine{}
+
+	if len(bootId) != 0 {
+		m.BootId = bootId
+	}
+
+	id, err := ioutil.ReadFile(boot_id_path)
+	if err != nil {
+		panic(err)
+	}
+	m.BootId = strings.TrimSpace(string(id))
+
+	return m
 }
